move serializers into separate module
